Add files via upload
This is my first java script project . In this project first i create a rectangle using HTML(5) and style this rectangle using CSS(3) and i give functionality using Java script. In this project whenever we hover the mouse pointer to the rectangle it change the color explore this project once you will love it and you can also use in your project to make project amazing.